+++ src/components/interactive/Chip.jsx
@@ -7,11 +7,17 @@ import baseElement from '../../util/baseElement';
 const Chip = ({ size, Link, theme, isActive, label, subtitle, onClose, ...rest }) => {
   const Element = baseElement({ href: rest.href, as: Link });
 
-  const rootClass = cx('fdsChip', 'border-focus', `fdsChip--${size}`, {
-    'fdsChip--active': isActive,
-    'fdsChip--blue': theme === 'blue',
-    'fdsChip--outline': theme === 'outline',
-  });
+  const rootClass = cx(
+    'fdsChip',
+    'border-focus--notrsn',
+    'trsn-default',
+    `fdsChip--${size}`,
+    {
+      'fdsChip--active': isActive,
+      'fdsChip--blue': theme === 'blue',
+      'fdsChip--outline': theme === 'outline',
+    }
+  );
   return (
     <Element {...rest} className={rootClass}>
       <span>